tempfile.rb: do not call File.identical? on closed stream
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@58794 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ lib/tempfile.rb
@@ -334,10 +334,12 @@ def Tempfile.create(basename="", tmpdir=nil, mode: 0, **options)
     begin
       yield tmpfile
     ensure
-      if File.identical?(tmpfile, tmpfile.path)
-        unlinked = File.unlink tmpfile.path rescue nil
+      unless tmpfile.closed?
+        if File.identical?(tmpfile, tmpfile.path)
+          unlinked = File.unlink tmpfile.path rescue nil
+        end
+        tmpfile.close
       end
-      tmpfile.close
       unless unlinked
         begin
           File.unlink tmpfile.path
